tidy up plugins
