+++ core/src/towers/mod.rs
@@ -1,3 +1,3 @@
 pub mod basic_tower;
 pub mod components;
-pub mod render_plugin;
+pub mod plugin;
