+++ core/src/units/mod.rs
@@ -1,4 +1,4 @@
 pub mod basic_unit;
 pub mod components;
-pub mod render_plugin;
-pub mod systems;
+pub mod plugin;
+mod systems;
